% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajiq_fxn.R
\name{calculate_and_project_likelihood_score}
\alias{calculate_and_project_likelihood_score}
\title{downsampleWith_group_by}
\usage{
calculate_and_project_likelihood_score(
  full_data,
  LM_data,
  LM_col,
  features,
  return_knn_graph = FALSE
)
}
\arguments{
\item{full_data}{dataset}

\item{LM_data}{data to use to cmopute likelihood scores (useful if you have a  very big dataset)}

\item{LM_col}{column containing your conditions (eg, 'condition' or 'treatment' or 'disease')}
}
\description{
calculate Likelihood score for group
}
