% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajiq_fxn.R
\name{differential_analysis_program}
\alias{differential_analysis_program}
\title{differential_analysis_program}
\usage{
differential_analysis_program(
  glm_input,
  outcome_features,
  contrast_variables,
  covariates_in_model = NULL,
  intercept = TRUE,
  contrast_method = "emm",
  SPLIT_BY_NAMES = NULL,
  return_conf_int_results = FALSE
)
}
\arguments{
\item{glm_input}{a dataframe of rows = cells or samples, and columns as your metadata and features of interest. These features should be outcome variables. Make sure column names are in R-friendly format. Be aware that + and - is not handled well in R. Clean your data by using gsub('\\+','pos',column_name) and gsub('\\+','neg',column_name) before using make.names(). Generally best to avoid any empty spaces.}

\item{outcome_features}{a vector of your outcome features - must have R-friendly format (use make.names() function). eg, c('B_cell_activated')}

\item{contrast_variables}{a vector of your contrast variables of interest. You likely just have just 1 value here,  unless you want to study interaction effects. eg, c('tissue')}

\item{covariates_in_model}{a vector of your covariates that you want to control for. Also used for a matched analysis. eg, c("mouse_id"),}

\item{intercept}{whether you want to set the intercept to 0 or not. Default is TRUE. Leave as-is unless you know to remove intercept.}

\item{contrast_method}{Either estimated marginal means or tukey ('emm','tukey'). defaulst to emm}

\item{SPLIT_BY_NAMES}{A vector of groups you want to restrict each analysis within. Use this if you want to subset the analyis within distinct groups (eg, within celltype), you can use this arguement to split the analysis. Eg ,c("celltype","tissue").}
}
\description{
wrapper function for differential analysis
}
