% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trajiq_fxn.R
\name{downsampleWith_group_by}
\alias{downsampleWith_group_by}
\title{downsampleWith_group_by}
\usage{
downsampleWith_group_by(df, sampleN)
}
\arguments{
\item{df}{dataset}

\item{sampleN}{the max # of observations (eg,cells) to sample per group. If}
}
\description{
downsampleWith_group_by will randomly sample from each group in your dataset that you've specified using group_by. If the # of observations is less than the desired # to sample, then it will sample all the cells in that group.
}
\details{
Authors: Meelad Amouzgar and David Glass
}
\author{
Meelad Amouzgar
date: June 13th, 2024
Ctrl+Shift+D namespace
usethis::use_pipe()
devtools::document()
}
